feat; add searchable dynamic index to fix deletion
